Added DocSamplingResult abstraction

+++ src/main/java/cc/mallet/topics/CollapsedLightLDA.java
@@ -906,8 +906,8 @@ public class CollapsedLightLDA extends ModifiedSimpleLDA implements LDAGibbsSamp
 		int [] tokenSequence = tokens.getFeatures();
 		int [] oneDocTopics = topics.getFeatures();
 
-		double[] localTopicCounts = new double[numTopics];
-		double[] localTopicCounts_i = new double[numTopics];
+		int[] localTopicCounts = new int[numTopics];
+		int[] localTopicCounts_i = new int[numTopics];
 		
 		// Populate topic counts
 		int nonZeroTopicCnt = 0; // Only needed for statistics
@@ -1047,7 +1047,7 @@ public class CollapsedLightLDA extends ModifiedSimpleLDA implements LDAGibbsSamp
 
 
 	static double calculateDocumentAcceptanceProbability(int[][] globalTypeTopicCounts, int[] globalTokensPerTopic,
-			double[] localTopicCounts, double[] localTopicCounts_i, int type, int oldTopic,
+			int[] localTopicCounts, int[] localTopicCounts_i, int type, int oldTopic,
 			int docTopicIndicatorProposal, double alpha, 
 			double beta, double betaSum) {
 		
@@ -1091,7 +1091,7 @@ public class CollapsedLightLDA extends ModifiedSimpleLDA implements LDAGibbsSamp
 
 
 	static double calculateWordAcceptanceProbability(int[][] globalTypeTopicCounts, int[] globalTokensPerTopic,
-			double[] localTopicCounts_i, int type, int oldTopic, int wordTopicIndicatorProposal, double alpha, 
+			int[] localTopicCounts_i, int type, int oldTopic, int wordTopicIndicatorProposal, double alpha, 
 			double beta, double betaSum) {
 		
 		// System.out.println("type: " + type + " oldTopic: " + oldTopic + " proposal: " + wordTopicIndicatorProposal);
